refactor(aggregator): split snapshotter module into multiple files
The snapshotter has become quite big with time with more than 900 lines
and multiples structs, this split makes the code more manageable.
